+++ src/chiabls/src/util.hpp
@@ -59,7 +59,7 @@ class Util {
     static std::string HexStr(const uint8_t* data, size_t len) {
         std::stringstream s;
         s << std::hex;
-        for (int i=0; i < len; ++i)
+        for (size_t i=0; i < len; ++i)
             s << std::setw(2) << std::setfill('0') << static_cast<int>(data[i]);
         return s.str();
     }
@@ -67,7 +67,7 @@ class Util {
     static std::string HexStr(const std::vector<uint8_t> &data) {
         std::stringstream s;
         s << std::hex;
-        for (int i=0; i < data.size(); ++i)
+        for (size_t i=0; i < data.size(); ++i)
             s << std::setw(2) << std::setfill('0') << static_cast<int>(data[i]);
         return s.str();
     }
